[API] Move csv logic out of useCase
