+++ api/lib/domain/usecases/index.js
@@ -173,7 +173,6 @@ module.exports = injectDependencies({
   getAnswer: require('./get-answer'),
   getAssessment: require('./get-assessment'),
   getAttendanceSheet: require('./get-attendance-sheet'),
-  getCandidatesImportSheet: require('./get-candidates-import-sheet'),
   getCampaign: require('./get-campaign'),
   getCampaignByCode: require('./get-campaign-by-code'),
   getCampaignAssessmentParticipation: require('./get-campaign-assessment-participation'),
@@ -209,6 +208,7 @@ module.exports = injectDependencies({
   getSessionCertificationReports: require('./get-session-certification-reports'),
   getSessionResults: require('./get-session-results'),
   getSessionResultsByResultRecipientEmail: require('./get-session-results-by-result-recipient-email'),
+  getSessionWithCandidates: require('./get-session-with-candidates'),
   getShareableCertificate: require('./certificate/get-shareable-certificate'),
   getTargetProfileDetails: require('./get-target-profile-details'),
   getUserByResetPasswordDemand: require('./get-user-by-reset-password-demand'),
